@PathVariable 변수 타입 변경
long으로 업캐스팅해서 쓰는 것이 나중에 더 용이 하기 때문에 long변수 타입으로 변경했습니다.

+++ src/main/java/com/codessquad/qna/domain/Question.java
@@ -1,16 +1,16 @@
 package com.codessquad.qna.domain;
 
 public class Question {
-    private int index;
+    private long index;
     private String writer;
     private String title;
     private String contents;
 
-    public void setIndex(int index) {
+    public void setIndex(long index) {
         this.index = index;
     }
 
-    public int getIndex() {
+    public long getIndex() {
         return index;
     }
